+++ src/Frontend/Dfe.Complete/Constants/ValidationConstants.cs
@@ -25,5 +25,6 @@
         public const string MustProvideDetails = "You must provide details";
         public const string MinisterNameRequired = "Enter the name of the minister that approved the decision";
         public const string DecisionDateRequired = "Enter a valid date the decision was made, like 27 3 2021";
+        public const string ValidDate = "Enter a valid date, like 1 1 2025";
     }
 }