+++ src/Frontend/Dfe.Complete/Pages/Projects/Decision/RecordDaoRevocation/MinisterName/AddDaoRevocationMinisterName.cshtml.cs
@@ -1,6 +1,6 @@
 using Dfe.Complete.Constants;
 using Dfe.Complete.Services;
-using DfE.CoreLibs.Caching.Interfaces;
+using GovUK.Dfe.CoreLibs.Caching.Interfaces;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using System.ComponentModel.DataAnnotations;