+++ src/Frontend/Dfe.Complete/Pages/Projects/Decision/RecordDaoRevocation/RecordDaoRevocation.cshtml.cs
@@ -1,5 +1,5 @@
 using Dfe.Complete.Constants;
-using DfE.CoreLibs.Caching.Interfaces;
+using GovUK.Dfe.CoreLibs.Caching.Interfaces;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;