+++ src/Tests/Dfe.Complete.CypressTests/cypress/e2e/projects/conversion/tasks/deed-of-variation.cy.ts
@@ -92,7 +92,7 @@ describe("Conversion tasks - Deed of variation", () => {
     });
 
     it("Should NOT see the 'save and return' button for another user's project", () => {
-        cy.visit(`projects/${otherUserProjectId}/tasks/handover`);
+        cy.visit(`projects/${otherUserProjectId}/tasks/deed_of_variation`);
         taskPage.noSaveAndReturnExists();
     });