+++ src/Tests/Dfe.Complete.Tests.Common/Customizations/Models/SignificantDateHistoryCustomization.cs
@@ -1,7 +1,7 @@
 ﻿using AutoFixture;
 using Dfe.Complete.Domain.Entities;
 using Dfe.Complete.Domain.ValueObjects;
-using DfE.CoreLibs.Testing.AutoFixture.Customizations;
+using GovUK.Dfe.CoreLibs.Testing.AutoFixture.Customizations;
 
 namespace Dfe.Complete.Tests.Common.Customizations.Models
 {