+++ src/Tests/Dfe.Complete.Tests/Constants/ValidationConstantsTests.cs
@@ -119,6 +119,11 @@ namespace Dfe.Complete.Tests.Constants
             Assert.Equal("Proposed capacity must be a number, like 345", ValidationConstants.ProposedCapacityMustBeNumber);
         }
         [Fact]
+        public void ValidDate_ShouldReturnExpectedMessage()
+        {
+            Assert.Equal("Enter a valid date, like 1 1 2025", ValidationConstants.ValidDate);
+        } 
+        [Fact]
         public void ChooseAtLeastOneReason_ShouldReturnExpectedMessage()
         {
             Assert.Equal("Select at least one reason", ValidationConstants.ChooseAtLeastOneReason);