+++ CHANGELOG.md
@@ -16,8 +16,26 @@ Security in case of vulnerabilities.
 
 ## Unreleased
 
+### Added
 
-See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-09-09.731...main) for everything awaiting release
+### Fixed
+- update "Task - Supplemental funding agreement - transfer - incorrect options"
+
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-09-12.746...main) for everything awaiting release
+
+---
+
+## [1.17.0](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2025-09-12.746) - 2025-09-12
+
+### Added
+
+- Added `Receive declaration of expenditure certificate` task for conversion and transfer projects.
+
+### Fixed
+- update "Give feedback about service" to use correct link
+- update /privacy and /accessibility to Allow Anonymous
+
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-09-09.731...production-2025-09-12.746) for everything in the release
 
 ---